fix(ui) refactor usage of the before and after components to clean up the node input submit button of specific use case logic
