+++ ui/components/NodeInputSubmit.tsx
@@ -2,7 +2,6 @@ import { getNodeLabel } from "@ory/integrations/ui";
 import { Button } from "@canonical/react-components";
 import { NodeInputProps } from "./helpers";
 import React, { Component, FC } from "react";
-import { useRouter } from "next/router";
 
 export const NodeInputSubmit: FC<NodeInputProps> = ({
   node,
@@ -11,8 +10,6 @@ export const NodeInputSubmit: FC<NodeInputProps> = ({
   disabled,
   dispatchSubmit,
 }) => {
-  const router = useRouter();
-
   const getProviderImage = (value: string) => {
     if (value.toLowerCase().startsWith("auth0")) {
       return "logos/Auth0.svg";
@@ -36,13 +33,6 @@ export const NodeInputSubmit: FC<NodeInputProps> = ({
   const isProvider = attributes.name === "provider";
   const provider = attributes.value as string;
   const image = getProviderImage(provider);
-  const showBackupLink =
-    (node.meta.label as unknown as { hasBackupLink: boolean })?.hasBackupLink ??
-    false;
-  const showTotpLink =
-    (node.meta.label as unknown as { hasTotpLink: boolean })?.hasTotpLink ??
-    false;
-  const showBackLink = node.meta?.label?.text === "Reset password";
 
   const getAppearance = () => {
     const appearance = (node.meta.label?.context as { appearance: string })
@@ -80,17 +70,6 @@ export const NodeInputSubmit: FC<NodeInputProps> = ({
   return (
     <>
       {beforeComponent}
-      {showBackLink && (
-        <Button
-          tabIndex={3}
-          type="button"
-          onClick={() => {
-            void window.history.back();
-          }}
-        >
-          Back
-        </Button>
-      )}
       <Button
         appearance={getAppearance()}
         tabIndex={4}
@@ -123,47 +102,6 @@ export const NodeInputSubmit: FC<NodeInputProps> = ({
           label
         )}
       </Button>
-      {showBackupLink && (
-        <Button
-          appearance="link"
-          tabIndex={5}
-          type="button"
-          onClick={() => {
-            const newQuery = { ...router.query, ["use_backup_code"]: "true" };
-            void router.push(
-              {
-                pathname: window.location.pathname,
-                query: newQuery,
-              },
-              undefined,
-              { shallow: true },
-            );
-          }}
-        >
-          Use backup code instead
-        </Button>
-      )}
-      {showTotpLink && (
-        <Button
-          appearance="link"
-          tabIndex={5}
-          type="button"
-          onClick={() => {
-            const newQuery = { ...router.query };
-            delete newQuery.use_backup_code;
-            void router.push(
-              {
-                pathname: window.location.pathname,
-                query: newQuery,
-              },
-              undefined,
-              { shallow: true },
-            );
-          }}
-        >
-          Use authentication code instead
-        </Button>
-      )}
       {afterComponent}
     </>
   );
